Remove migrator class
Test the `Nicknamizable` concern instead.
