+++ decidim-core/db/migrate/20171212103803_create_unique_nicknames.rb
@@ -1,15 +1,23 @@
 # frozen_string_literal: true
 
-require "decidim/migrators/username_to_nickname"
-
 class CreateUniqueNicknames < ActiveRecord::Migration[5.1]
-  def change
+  class User < ApplicationRecord
+    extend Decidim::Nicknamizable
+
+    self.table_name = :decidim_users
+  end
+
+  def up
     add_column :decidim_users, :nickname, :string, limit: 20
 
-    reversible do |direction|
-      direction.up { Decidim::Migrators::UsernameToNickname.new.migrate! }
+    User.find_each do |user|
+      user.update!(nickname: User.nicknamize(user.name))
     end
 
     add_index :decidim_users, :nickname, unique: true
+  end
+
+  def down
+    remove_column :decidim_users, :nickname
   end
 end
